feat: sécurité pour lédition.

+++ src/Controller/ArticleController.php
@@ -3,6 +3,7 @@
 namespace App\Controller;
 
 use App\Entity\Article;
+use App\Entity\User;
 use App\Form\ArticleType;
 use App\Repository\ArticleRepository;
 use Doctrine\ORM\EntityManagerInterface;
@@ -10,6 +11,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 use Symfony\Component\Security\Http\Attribute\IsGranted;
 
 #[Route('/article')]
@@ -32,14 +34,16 @@ class ArticleController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $article->setAuthor($this->getUser());
+            /** @var User $user */
+            $user = $this->getUser();
+            $article->setAuthor($user);
             $entityManager->persist($article);
             $entityManager->flush();
 
             return $this->redirectToRoute('article_index', [], Response::HTTP_SEE_OTHER);
         }
 
-        return $this->renderForm('article/new.html.twig', [
+        return $this->render('article/new.html.twig', [
             'article' => $article,
             'form' => $form,
         ]);
@@ -56,6 +60,18 @@ class ArticleController extends AbstractController
     #[Route('/{id}/edit', name: 'article_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Article $article, EntityManagerInterface $entityManager): Response
     {
+        /** @var User|null $currentUser */
+        $currentUser = $this->getUser();
+
+        $isAuthor = $article->getAuthor()?->getId() === $currentUser?->getId();
+        $isAdmin = $this->isGranted('ROLE_ADMIN');
+
+        if (!$isAuthor && !$isAdmin) {
+            $this->addFlash('danger', "Vous n'êtes pas autorisé à éditer cet article.");
+
+            return $this->redirectToRoute('article_index');
+        }
+
         $form = $this->createForm(ArticleType::class, $article);
         $form->handleRequest($request);
 
@@ -65,7 +81,7 @@ class ArticleController extends AbstractController
             return $this->redirectToRoute('article_index', [], Response::HTTP_SEE_OTHER);
         }
 
-        return $this->renderForm('article/edit.html.twig', [
+        return $this->render('article/edit.html.twig', [
             'article' => $article,
             'form' => $form,
         ]);
@@ -74,6 +90,18 @@ class ArticleController extends AbstractController
     #[Route('/{id}', name: 'article_delete', methods: ['POST'])]
     public function delete(Request $request, Article $article, EntityManagerInterface $entityManager): Response
     {
+        /** @var User|null $currentUser */
+        $currentUser = $this->getUser();
+
+        $isAuthor = $article->getAuthor()?->getId() === $currentUser?->getId();
+        $isAdmin = $this->isGranted('ROLE_ADMIN');
+
+        if (!$isAuthor && !$isAdmin) {
+            $this->addFlash('danger', "Vous n'êtes pas autorisé à supprimer cet article.");
+
+            return $this->redirectToRoute('article_index');
+        }
+
         if ($this->isCsrfTokenValid('delete' . $article->getId(), $request->request->get('_token'))) {
             $entityManager->remove($article);
             $entityManager->flush();
